-Fixed bug where you cant scan codes from camera photos -Moved DecodeBarcode to helper
